스위치 작동 확인 하지만 버그 발견 - 엔디안 오류가 있는 것 같음

+++ Sixtar-Switch_RP2040/src/backup/SwitchControlLibrary.h
@@ -2,20 +2,20 @@
 
 struct Button
 {
-    static const uint16_t Y = 0x0001;
-    static const uint16_t B = 0x0002;
-    static const uint16_t A = 0x0004;
-    static const uint16_t X = 0x0008;
-    static const uint16_t L = 0x0010;
-    static const uint16_t R = 0x0020;
-    static const uint16_t ZL = 0x0040;
-    static const uint16_t ZR = 0x0080;
-    static const uint16_t MINUS = 0x0100;
-    static const uint16_t PLUS = 0x0200;
-    static const uint16_t LCLICK = 0x0400;
-    static const uint16_t RCLICK = 0x0800;
-    static const uint16_t HOME = 0x1000;
-    static const uint16_t CAPTURE = 0x2000;
+    static const uint8_t Y = 0x0001; // minus?
+    static const uint8_t B = 0x0002; // plus?
+    static const uint8_t A = 0x0004; // lclick?
+    static const uint8_t X = 0x0008; // rclick?
+    static const uint8_t L = 0x0010; // home?
+    static const uint8_t R = 0x0020; // capture?
+    static const uint8_t ZL = 0x0040;
+    static const uint8_t ZR = 0x0080;
+    static const uint8_t MINUS = 0x0100;
+    static const uint8_t PLUS = 0x0200;
+    static const uint8_t LCLICK = 0x0400;
+    static const uint8_t RCLICK = 0x0800;
+    static const uint8_t HOME = 0x1000;
+    static const uint8_t CAPTURE = 0x2000;
 };
 
 struct Hat
